backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# PDF Processing
PyPDF2==3.0.1

# AI/ML
groq>=0.11.0

# Database
sqlalchemy>=2.0.25

# Data handling
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0
